closes #15, closes #16 Need a letter
Added the validation that the password needs at least one letter
character and the associated configuration. Also refactored the building
of the validator object itself so that the settings are removed to a
separate class. The settings class and configuration section share the
same interface so that both can be used interchangeably.