More Prep for #22
The Needs Number validation has been refactored into a Handler. The
service locator has been updated to chain the Needs Number validator to
the Minimum Length validator. The Validator has removed the specific
needs number validation code (as it is now in the handler) and the
return type has been changed to a bool (as there is no indication which
handler a failure came from, and in retrospect it probably isn't that
important. What is important is that the pasword passed or failed the
validation as a whole)