Prep for #13 Strong name the assemblies
Strong name the assemblies.
NOTE: Only the public key is included in the repository. In order to use
the assemblies produced by a build of the project you will have to run
the following command to skip the verification process:

sn -Vr *,fe72000dffcf195f

It is not recommended to do this in a production environment.