Renamed a property in IPasswordValidationSettings
It made more sense for CustomSettings to be called CustomData because it
may not necessarily be settings that are passed in but general data that
allows the custom validation handler to operate. In fact, since any
object may be passed in (including objects that can query persistance
stores, etc. CustomData may not be the best name either)